requirements.txt
langchain
langchain-openai
langchain-community
chromadb
python-dotenv
openai
streamlit

